+++ package.json
@@ -66,6 +66,7 @@
 		"husky": "^8.0.3",
 		"jest": "^29.7.0",
 		"jest-environment-jsdom": "^29.7.0",
+		"jsdoc": "^4.0.2",
 		"lint-staged": "^15.2.0",
 		"msw": "^2.0.11",
 		"styled-components": "^6.1.3",
