+++ src/component/common/BottomSheet/BottomSheet.tsx
@@ -1,7 +1,15 @@
 import React from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
 
-interface bottomSheetProps {
+/**
+ * @description 바텀시트 공통 컴포넌트입니다.
+ * @param {isOpen} useState로 관리해주는 열고 닫는 boolean 값입니다.
+ * @param {onClose} 닫는 함수입니다.
+ * @param {viewHeight} 얼만큼 높이로 바텀시트를 보여줄 지 설정할 수 있습니다. string ex) calc(100vh * 0.2)
+ * @param {children} 바텀 시트 안에 새로운 컴포넌트 혹은 내용을 넣어줄 수 있습니다.
+ */
+
+interface BottomSheetProps {
 	isOpen: boolean;
 	onClose: () => void;
 	viewHeight: string;
@@ -13,7 +21,7 @@ const BottomSheet = ({
 	onClose,
 	viewHeight,
 	children,
-}: bottomSheetProps) => {
+}: BottomSheetProps) => {
 	return (
 		<AnimatePresence>
 			{isOpen && (
